refactor: DummyData 분리
